+++ CHANGELOG.md
@@ -4,6 +4,14 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.4.0](https://github.com/nozaq/terraform-aws-lambda-auto-package/compare/v0.3.0...v0.4.0) (2022-12-08)
+
+
+### Features
+
+* **iam:** Add permissions_boundary variable ([#26](https://github.com/nozaq/terraform-aws-lambda-auto-package/issues/26)) ([2f1cc73](https://github.com/nozaq/terraform-aws-lambda-auto-package/commit/2f1cc73bef98b05cbed65557e5c91af286aae2f4))
+* **lambda:** Add lambda_kms_key_arn variable ([#24](https://github.com/nozaq/terraform-aws-lambda-auto-package/issues/24)) ([2575a45](https://github.com/nozaq/terraform-aws-lambda-auto-package/commit/2575a45fd44b6e33f3b5908d53f3b7fef3704563))
+
 ## [0.3.0] - 2022-02-21
 
 ## [0.2.1] - 2021-11-21